scripts.logwatcher: Consider more log lines as progress
Having a large number of buildslaves caused the start of the master
to be considered a failure as attaching individual buildslaves was
not considered log progress and thus the start timed out.

Attaches are now considered part of log progress and the timeout
has been increased to better cope with the extra load of having
a lot of buildslaves.

Furthermore, paused slaves will not be shut down anymore on restart
as their state should be persisted and restored after the restart.