Placing artifacts into one folder per builder
This should help alleviate filesystem operation at the artifact server root, since we should have way less folders in the same directory

This commits sets August 1st 2017 as the change date (everything upload after that will be put in the new path).
This commit must be deployed to all servers before this date, so it can be pushed further into the future if needed.

CDS-974